Added Klingon lang added oggplayer (optimised & stripped of pointless stuff) add URI loader
(debug code when pressing 1 on intro - removedfor release)

+++ BoltThrower/source/LWOSource/lwo2.cpp
@@ -165,7 +165,7 @@ lwObject *lwGetObject( const char* filename, unsigned int *failID, int *failpos 
 
             rlen = get_flen();
             if ( rlen < 0 || rlen > (int)cksize ) goto Fail;
-            if ( rlen <= cksize - 2 )
+            if ( rlen <= (int)cksize - 2 )
                layer->parent = (int)getU2( fp );
             rlen = get_flen();
             if ( rlen < (int)cksize )
